requirements.txt
Django==5.2.12
djangorestframework==3.15.2
django-filter==25.2
django-cors-headers==4.9.0
graphene-django==3.2.3
graphene==3.4.3
psycopg2-binary==2.9.11
requests==2.32.3
python-dotenv==1.0.1
ollama==0.4.7
gunicorn==23.0.0